data-collection/requirements.txt
requests~=2.32.3
flask~=3.0.3
pymongo~=4.7.3

google~=3.0.0
ipython~=8.18.1

python-dotenv~=1.0.1

marshmallow~=3.21.3
